update: improve qx_registry implementation and add tests
