+++ python/qx_utilities/gmri
@@ -35,7 +35,7 @@ from qx_utilities.general import matlab as gm
 from qx_utilities.general import core as gc
 from qx_utilities.general import exceptions as ge
 from qx_utilities.general import commands_support as gcs
-from qx_registry import build_qx_registry, load_qx_registry, qx_commands
+from qx_registry import qx_commands
 
 
 help = r"""DESCRIPTION: QuNex suite python-based general neuroimaging utilities
@@ -437,6 +437,7 @@ def main(args=None):
     try:
         if comm == 'build_qx_registry':
             print("Building command registry...")
+            from qx_registry_build import build_qx_registry
             build_qx_registry()
             print("Command registry successfully built!")
             sys.exit(0)
